requirements-dev.txt
# coveragepy==5.0 fails with `Safety level may not be changed inside a transaction
# on python 3.6.0 (xenial)
pip>=21.1
urllib3>=1.26.4
coverage>=6.0b1
flake8
mock
monotonic
pre-commit
pytest
pyyaml
responses
tox-pip-extensions
tox>=3.8
unidiff
responses
ibm_db
boxsdk[jwt]
pyahocorasick
safety
requests==2.25.0
packaging==20.7
